Assignment7/requirements.txt
geonames_rdf
